ci-cd: move composingBootstrap to ci-cd:common
