+++ generators/ci-cd/generators/common/generator.ts
@@ -34,6 +34,21 @@ export default class CommonGenerator extends BaseSimpleApplicationGenerator<CiCd
     }
   }
 
+  get composingBootstrap() {
+    return this.asComposingBootstrapTaskGroup({
+      async composingBootstrap({ application }) {
+        const { backendType = 'Java' } = application;
+        if (['Java', 'SpringBoot'].includes(backendType)) {
+          await this.dependsOnBootstrap('java');
+        }
+      },
+    });
+  }
+
+  get [BaseSimpleApplicationGenerator.COMPOSING_BOOTSTRAP]() {
+    return this.delegateTasksToBlueprint(() => this.composingBootstrap);
+  }
+
   get preparing() {
     return this.asPreparingTaskGroup({
       preparing({ applicationDefaults }) {
